% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg class}
\value{
none
}
\description{
creates multiple linear regression model
}
\section{Fields}{

\describe{
\item{\code{regcoeff}}{regression coefficients}

\item{\code{fitted_value}}{for fitted values}

\item{\code{df_name}}{for input dataset name}

\item{\code{residu}}{for residuals}

\item{\code{degree_of_freedom}}{for degree of freedom}

\item{\code{residual_variance}}{for residual variance}

\item{\code{var_reg_coeff}}{for variance of regression coefficients}

\item{\code{t_value}}{for t value}

\item{\code{pt_value}}{for p value}
}}


